Book: some mandatory changes in building the extended PSF section
Until now, there were some mistakes in the tuotrial of the making the
extended PSF, for instance name of 'astcrop' input was incorrect, the
kernel directory did not determinesd, the word how was repetead twice in
the text.

With this commit, the input name of 'astcrop' has been corrected, the
directory of the kernel has been determined and finally the extra word are
deleted.