Book: Full edit of saturated pixel section of PSF tutorial
Until now, in the "Saturated pixels and Segment’s clumps" section of the
extended PSF tutorial, we were running NoiseChisel on the same image as the
convolved image (by mistake: the two files were the same!). As a result,
any future measurement on the Segment output would be severely wrong for
the saturated objects. This was because the saturated pixels have been
filled with the maximum of their nearby pixels. Furthermore, no description
had been given on how to identify objects/clumps affected by saturation.

With this commit, the section has been edited and clarified to make a new
Segment output where the saturated pixels are blank. An example MakeCatalog
command has also been given to show how we can find the clumps affected by
saturation.