PSF-stamp: normalization value fixed to float32
Until now, the numerical type of the stamp images could be different for
stamps generated by the same image! This was because the type of the
normalization value are interpretted from the number of digits they
have. This could cause problems later when we stack the images. The issue
(with examples) is fully described in the comments that have been added to
the scipt as part of this commit.

With this commit, at first the type of the normalization value is changed
to the 'float32' and then images are normalized with this simillar type of
normalization values. Finally the type of the stamp images will be same and
we can easily by using the 'astarithmetic' stack all of the stamp images
together.