Library (wcs.h): wcsset is run after setting PC+CDELT
Until now, within 'gal_wcs_decompose_pc_cdelt' we would just update the
'wcs->pc' and 'wcs->cdelt' values, but we weren't running 'wcsset'. As a
result, WCSLIB was using some internal values prior to our change. This
would result in in-correct outputs when the original rotation was defined
by CROTA (which has been depreciated). For example in the following set of
commands:

    # Download one (problematic!) dataset:
    raw=NGA_NGC1052-nd-int.fits.gz
    url1=http://galex.stsci.edu/data/GR6/pipe/01-vsn
    url2=05038-NGA_NGC1052/d/01-main/0001-img/07-try
    wget $url1/$url2/$img

    # Convert RA/Dec to X/Y (will be wrong!)
    echo 40.097200 -8.432200 \
         | asttable -c'arith $1 $2 wcs-to-img' \
                    --wcsfile=$img --wcshdu=0

    # Re-create the whole image (including WCS):
    astcrop $img -h0 --mode=img --section=:,: --output=good.fits

    # Convert RA/Dec to X/Y (will be correct!)
    echo 40.097200 -8.432200 \
         | asttable -c'arith $1 $2 wcs-to-img' \
                    --wcsfile=good.fits  --wcshdu=1

After investigating the commands above, I noticed that the reason it gives
a correct value on 'good.fits' is that we haven't been running 'wcssset'
after correcting the internal structure information of the WCSLIB
structure.

With this commit, a call to 'wcsset' has been added after modifying the
internal values in 'gal_wcs_decompose_pc_cdelt' and this fixed the problem.
I also made two other minor edits:

 - In the book, there was no space added after the function name of
   'gal_wcs_write_in_fitsptr' (which is not the standard way we do it for
   all functions).

 - The error messages for 'wcsset' are now more standard (printing the name
   of the program, followed by function name, and then the WCSLIB message).

This bug was reported by Ignacio Ruiz Cejudo and Raul Infante-Sainz.

This fixes bug #62052.