Table: --catrowfile will not free string column contents
Until now, when Table's row concatenation ('--catrowfile') was used with a
column containing strings, it would mistakenly free the contents of the
strings. Therefore when we later wanted to write the string column, Table
would crash with a segmentation fault!

With this commit, after copying the pointers to the strings from the input
tables into the output table, we set the initial points to 'NULL' so they
aren't freed. This fixes the problem.

I also noticed that I had forgot to add Ignacio Ruiz Cejudo to the
'doc/announce-acknowledge.txt' for his help in the previous bug.

This bug was reported by Manuel Sánchez-Benavente.

This fixes bug #62054.