NEWS: minor edit and typo corrections
Until now, the notices for the not-yet-released version of Gnuastro had a
'[stable]' on the title which could be confusing. Also, there were some
minor typos within the NEWS file.

With this commit, square brackets after the latest version say '[not yet
released]', and a spell-check was run to find and fix typos.

I also noticed that bug #61976 was listed before bug #61967, so to be more
clear (and follow the bug code sorting), it has been placed after it.