Library (type.h): signed ints, same width preferred in reading strings
Until now, when reading a number from a string if it was positive, we would
only assign it to an un-signed integer type: the number of bits in the
integer was determined based on its largest possible value. However, in C's
internal automatic type conversion (when dealing with binary operators like
'+' or '>'), when an un-signed and signed integer of same width is given to
a binary operator, they are both converted to the unsigned type.

This would produce unexpected results! Because a signed integer can have a
value like '-1'. But when this value is converted to an unsigned integer of
the same width, it becomes the maximum value! This issue can be
demonstrated with the commands below:

  astarithmetic 200 200 2 makenew 5 mknoise-sigma int32 \
                --output=int-noise.fits
  astarithmetic int-noise.fits 100000 gt --output=thresh.fits

The standard deviation of 'int-noise.fits' is 100, so its maximum will be
around 500 (5-sigma). Therefore, none of the pixels should be larger than
100000 and 'thresh.fits' should only have a value of 0. However, until this
commit 'thresh.fits' would have many '1'-valued pixels.

With this commit, this problem has been addressed in three levels:

 - In 'gal_type_string_to_number' (where we read a number from the
   command-line string that the user has given), positive values aren't
   only given to unsigned types. If a positive value is smaller than the
   largest possible signed value, it will be given a signed type.

 - In 'arithmetic_binary', a new sanity check has been added: when the
   inputs have the same width and are integers, but with different signs, a
   warning will be printed to fully elaborate the problem to the user and
   propose solutions.

This bug was reported by Zohreh Ghaffari and Pedram Ashofteh Ardakani.

This fixes bug #62069.