Table: properly freeing non-used string row contents with --tail
Until now, when the value to '--tail' was more than half the number of rows
and the output columns contained strings, Table would crash with a
segmentation fault. This was because of an incorrect freeing of string
column arrays for the un-used columns.

With this commit, we now properly account for cases where the number of
tail rows is more than half of the total number of rows and only free those
that will not be in the output.

This bug was reported by Manuel Sánchez-Benavente.

This fixes bug #62070.