astscript-radial-profile: including the --azimuth option
Until this commit, the radial profile script was not able to generate
azimuthal radial profiles. The reason is that the azimuthal profile of
MakeProfile was added after the radial profile script was included in
Gnuastro.

With this commit, the necessary steps have been added in order to be able
to construct azimuthal radial profiles. The option --azimuth (or -a) is
used for specifying two angles that will be considered together with the
radial apertures to later measure the radial profile only in such portion
of the image.