Webpage: Installation section now has Conda SVG and link
Until now, the webpage didn't have any mention of the fact that Gnuastro is
also packaged in Conda. Since many people use it regularly and are
comfortable with Conda, its useful for many potential new users to be able
to easily test it.

With this commit, the first paragraph of the "Installation" section in the
main wepage first mentions that Gnuastro is available in OS package
managers, then it mentions Conda, and finally it has a link to the "Quick
start" page that describes building from source.