Radial-profile: Remove symbolic link directory from the temporary directory
Until now, when the given temporary directory was relative, the script
would crash. Because we had used a symbolic link for saving some outputs
that did not change. In this situation the script would search for the
relative temporary directory in the temporary directory, which does not
exist, so it would crash.

With this commit, everywhere that we want to use the symbolic links, first
we change to the temporary directory and then make our symbolic link
without any directory information. After making our symbolic link, we
change back to the runnnig directory. This is necessary because the
'realpath' command does not exist in some operating systems.