fits-view: new script for visulaization the multi images and table
Until now, for opening the multi images together and then match and lock
together we had to use the command line and press many bottons, and if you
want to open a table with topcat you have to write it in the command line.

With this commit, the script in the "other useful software" appendix has
been brought in to the 'bin/script' to be later released as an
'astscript-*'.