fits-view: now gets installed with its .desktop file, doc added
Until now, the 'astscript-fits-view' installed script that was added in the
previous commit, was just the bare script, without any documentation,
option-parsing, any accompanying '.desktop' file, and finally, it wasn't
installed!

With this commit, all these components have been added to the script, and
it now gets installed with all the other installed scripts. A new
'.desktop' file has also been added for it, which gets installed in the
'PREFIX/share/gnuastro' directory, enabling users to easily link with it in
their '~/.local/share/applications'. The instructions to install the
'.desktop' file are now printed at the end of 'make install'.

A documentation has also been added in the book for this script, and to be
complete, a new appendix has been added on how to install TOPCAT.