Radial-profile: using symbolic link when no --azimuth option is used
Until now, when the option --azimuth was not invoked, the script just
asumed that the apertures image was the radial apertures image. When the
option --azimuth was invoked, a combination of azimuth and radial apertures
was computed to finally make the measurements. This was working fine.
However, in order to follow the same conventions as in other parts of the
code, it is good to explicitly create symbolic links to those images that
are not necessary. The reason is that it is clearer and easy to find bugs.

With this commit, a symbolic link is generated when the option --azimuth is
not used. By doing this, the final aperture image (those that is used for
making the measurements) is directly set to be the radial apertures image.
If the user wants to make an azimuth radial profile, then a combination of
radial and azimuth apertures is constructed. In the Arithmetic command, I
have forced the 'arc' image to be of 'int16' type to properly account it.

In addition to this, several parts of the code have been changed to include
better information and follow conventions.