Book: considering only two set of stars for the extended PSF tutorial
Until now, the range of magnitudes for constructing the outer and inner
part of the PSF (in the Building the extended PSF tutorial) were such as we
only had a very few images for the outer part, and also saturated stars for
the inner part. Although the original idea was to construct several inner
parts and concatenate them together to have high signal to noise in the
outer part and not saturation in the center, that could make the tutorial
very long and difficult to follow. As a consequence, I decided to simplify
it to only have two PSF parts: outer and inner.

With this commit, only two set of stars are considered. One for the outer
part (although mixing very different brightness stars) and the other for
the core of the PSF (mixing saturated and not saturated stars). Eventhough
this would not be the reality, it simplify a lot the tutorial because only
two different PSFs are necessary while explaining the main ideas.

With this commit several parts of the tutorial have been changed to take
into account this changes. In addition to this, minor typos have been
corrected. Right now, everything is ready to continue describing the
process of joining the two different parts of the PSF (outer and inner).