Book: modifying the section about the subtraction of the extended PSF
Until now, the description on how to make the subtraction of the extended
PSF was not up to date (taking into account the changes in the previous
sections/commits).

With this commit, this section has been modified in order to continue the
tutorial and make the subtraction. As a consequence, right now the tutorial
has been completed.

However, there are a couple of remaining things: 1.  Obtain better
segmentation object image and 2. Ensure that the PSF is good enough so it
does not leave strong residuals in the subtracted image. In addition to
that, a final revision for polishing and correcting typos is needed.