NoiseChisel: freed tile-checking pointer set to NULL
Until now, we were simply freeing the 'tilecheckname' element of the
tessellation structure (which is used with the '--checktiles' option) after
it wasn't necessary, but not setting it to NULL. As a result, when
'--checktiles' was called with '--continueaftercheck', NoiseChisel would
crash.

With this commit, after freeing the pointer, it is set to NULL and this
avoids any segmentation fault at freeing time.

This bug was reported by Giulia Golini.

This fixes bug #62112.