Book: optimizing segmentation objects for extended PSF construction
Until now, the object extension we were obtaining in the tutorial of the
extended PSF construction was not so good. The reason is that the bright
stars were broken into different objects. As a consequence, the outer part
of these stars were masked and the estimation of these regions was poor.

With this commit, a small paragraph has been added in order to explain how
to solve this problem. It basically consists in decrease two parameters of
Segment: --gthresh and --objbordersn. By decreasing them, the bright stars
are not broken into different objects (specially in the outer part). In
addition to this, the magnitude range of the bright stars has been
decreased in order to consider brighter stars and not fainter ones.