PSF scripts: forcing to use float32 operations with Arithmetic
Until now, the operations done with Arithmetic were done without explicitly
setting the data type. However, as a consequence of a recent change, in
some cases that causes a problem. The issue is described in bug #62096:
https://savannah.gnu.org/bugs/?62096

With this commit, this issue has been fixed for the PSF scripts. In all
places where Arithmetic is used and it is possible that the data type are
uint or int, they are forced to be float32. By doing this, the scripts work
without any problem.