PSF scripts: adding the --segment option for compuing the flux factor
Until now, the computation of the flux was done without using the
segmentation image. Indeed, two PSF scripts had deprecated options related
with the masking of contaminant sources.

With this commit, these options have been changed in order to use the
option --segment, that takes as an input the Segment's output to mask non
wanted sources. In the same way, the necessary parts of the tutorial and
descriptions of these options in the book have been added.