Book: new section on integer benefits and pitfalls
Until now, there was no good explanation on the how to effectively use
integers to benefit mostly from their improved speed and less storage,
while avoiding some common pitfalls.

With this commit, a new sub-section has been added under the Arithmetic
section that explains these benefits and pitfalls.

Furthermore, following the recent change in Commit 213a1976 (reading
integers from the command-line preferrably into their signed format, when
they fit the range), the ordering of integer types was changed so signed
integers of same width have a lower identification code. This was necessary
because we select the output type of binary operations from this type code
identifier. This change fixed the unreasonable output of bug #62096:

   astarithmetic 250 1 +

But "strange" (for the un-expert) outputs can still happen like the example
given the new section of the book ('125 10 +'). So hopefully the
explanation given in that section will avoid confusions.

While fixing this condition, I noticed that the warning for equal-width but
differently-signed integers was not getting suppressed with the '--quiet'
option! So that issue has also been fixed with this commit.

This bug was reported by Raul Infante-Sainz.

This fixes bug #62096.