fits-view: only using resolution of primary image
Until now, 'astscript-fits-view' would simply read all the active screen
resolutions (those with a '*'). However, when more than one display is
connected to the computer, there will be multiple lines with a '*' in the
output of 'xrandr'!

With this commit, only the first line with a '*' is used (which is the
primary display).