Installation: fix GSL Steffen interpolation bug
Until now, Gnuastro compiled the interpolation library based on the GSL
Steffen interpolation without a proper check flag in the configuration
files. This resulted in the following error message when one tried to use
the 'gal_interpolate_1d_make_gsl_spline' function with the
'GAL_INTERPOLATE_1D_STEFFEN' option:

     gal_interpolate_1d_make_gsl_spline: Steffen interpolation isn't
     available in the system's GNU Scientific Library (GSL). Please install
     a more recent GSL (version >= 2.0, released in October 2015) and
     rebuild Gnuastro

With this commit, we have defined the 'GAL_CONFIG_HAVE_GSL_INTERP_STEFFEN'
and 'HAVE_GSL_STEFFEN' flags in the configuration and library Makefiles, to
check if the host machine can handle the Steffen interpolation from GSL.

Also, the interpolation example in the Gnuastro manual has been updated to
reflect changes on the 'gal_interpolate_1d_make_gsl_spline'
arguments. Additionally, a sample 'table.txt' file was added to the
example, and the build step using the 'astbuildprog' program was
demonstrated too.

This fixes bug#62127