psf-select-stars: added missing symbolic link to astmatch in tests/
Until now, in the test of the 'psf-create-select-stars' we did not create a
symbolic link to 'astmatch'. Therefore, in the 'make check' step the script
search for 'astmatch' in the host operating system.

With this commit, a symbolic link to the internaly built 'astmatch' is
created by the testing script. This fixed this problem.