Script checks: all built Gnuastro programs now in devoted dir
Until now, during 'make check', each of the installed scripts would re-make
symbolic links to the Gnuastro programs it needed into the build
directory's 'tests/' directory (mixed with all the data files that are
created there during the other tests!). This had several problems:

 - Forgetting to add one program in each test script of an installed script
   would cause a failure.

 - The various installed script's tests would over-write the symbolic links
   when run in parallell, potentially causing corrupt files.

With this commit, the 'prepconf.sh' script (which is the first test script
that is run with 'make check', and designed initially for a similar problem
with configuration files) puts a symbolic link to all built programs and
installed scripts in a specific directory called 'tests/programs-built' (in
the build directory). That directory is also added to the 'PATH' of all
installed script tests before they are run; ensuring that the currently
built versions of the programs are run, not the ones that may (or may not!)
exist on the host operating system.

Also, in the process, the following changes were made in the scripts:

 - In all installed scripts, the 'on_off_option_error' doesn't end the
   error messages with a '.' (this is part of the GNU Coding Standards:
   error messages shouldn't end in a point).

 - In all installed scripts, the 'check_v' function error message is a
   single line, not two 'echo's. It is important that error messages be
   printed in one line, otherwise, they will get mixed with other messages
   when run in parallel.

 - In 'astscript-psf-create-make-stamp', the error messages are now in one
   long line instead of multiple 'echo's, and the wrong option values are
   also printed in the message. This will help debugging.