Tests: LANG=C is set avoid floating points with , in some systems
Until now, the outputs of some shell number substitutions would be printed
in the local language of the host operating system. This would cause
problems in languages that use a coma (,) instead of a point (.) for the
floating point of numbers. This problem was introduced in the recently
added tests for 'astscript-psf-create-make-stamp' and
'astscript-psf-model-flux-factor' (where we read the center of the image
and give it as a value in the 'make check' tests).

With this commit, the 'tests/Makefile.am' now also exports 'LANG=C' to make
sure that all the tests are run in the standard C environment, allowing the
scripts to print the numbers with a (.) as floating point.

This issue was found with the help of Manuel Sánchez-Benavente.