Installed scripts: LANG=C internally set in all scripts
Until now, when passing values from command-line tools like AWK or SED, the
operation would depend on the language of the host operating system. In
particular, floating point numbers that are printed like '23,45' in some
languages (as opposed to the expected '23.45').

With this commit, at the very start of all the installed scripts we now
have an 'export LANG=C' to ensure that problems like above don't happen.