PSF-scripts: completed extended PSF tutorial
Until now, the final part of the PSF subtraction (last!) section of the
tutorial was incomplete. Primarily because after the subtraction of the
brightest star, we would get a very noisy region which corresponded to the
PSF footprint.

After some inspection, I found the result and have fully explained it in
the updated text of that tutorial. In short, it happens because of the low
S/N of the outer parts. The solution is to only subtract stars that have a
scaling factor less than the improved S/N of the stacked PSF.

In the process, two new options were added to the 'astscript-fits-view'
script to help setting the "Pan" of DS9 from the command-line (where we can
have the coordinates as a shell variable in Gnuastro format: with a comma
in the middle).