PSF-scripts: added info in NEWS, README and list of programs in book
Until now, these scripts were only mentioned in the book, and not the most
visible parts of Gnuastro (the 'NEWS', 'README', or the appendix with the
list of programs). This was because work on them hadn't reached a first
level of completion yet.

With this commit, they have been added to those parts of Gnuastro.
Especially because the tutorial devoted to their usage demonstration has
been completed. Of course, I am sure many problems may be present or we may
find many bugs, but they are now sufficiently tested and documented for
first public release.