Book: better demonstration of the PSF-subtracted image
Until now, we only asked the user to manually zoom-in to one of the bright
stars to see the effect of subtraction.

With this commit a complete 'astscript-fits-view' file has been given to
immediately zoom into the brightest star that was actually subtracted. To
do this, it was also necessary to add the creation of a log-file to the
subtraction script.