Installed scripts: not calling bash, several stylistic improvements
Until now, the 'astscript-fits-view' installed script had a shebang of
'/bin/bash', but this is not portable (not all systems have Bash!). All
other scripts had a standard '/bin/sh' shebang.

Also, to check equality in some parts of that script we were using '=='
instead of '=' to check for equality. But this doesn't follow stylistic
conventions of all our other installed scripts!

Also, when needing to check a variable against more than one string, the
scripts would use '[ $abc = x ] || [ $abc = y ]'. This is against good
shell scripting conventions because it involves two invocations of '[' (or
the 'test' program). However, if we simply use '-o' between the checks
(like '[ $abc = x -o $abc = y ]'), then we only call 'test' once and thus
speed up the script.

Finally, in some places, we were quoting static strings like '[ "$mode" =
"wcs"]'. This is also against good stylistc conventions in shell scripting:
static strings don't need quotation, so adding quotations over them will
confuse any reader.

With this commit, all the issues above have been addressed and these issues
have been uniformalized in the installed scripts. We should later define
these as coding conventions for installed scripts.

The points above were all raised [1] by Alexey Dokuchaev (who is kindly
packaging Gnuastro in FreeBSD). The point on the redundancy of "x-escaping"
(or using things like '[ x"$hdu" = x ]'), was not implemented because I see
it used in the './configure' script and have bad memories of not using it
(now that I checked, they may have been due to not quoting the shell
variable!).

While inspecting the codes, I noticed that when adding multiple input file
arguments to the scripts (during the 'while' script where we parse the
command-line arguments and options), when multiple arguments were given, we
were mistakenly reversing their order and adding an extra space to the
input argument name (with 'inputs="$1 $inputs"'). So the relevant line in
all the scripts was modified to put 'inputs="$1"' when 'inputs' is empty,
and 'inputs="$inputs $1"' when 'inputs' already has a value (more than one
argument was given). This fixes the order, and also doesn't add an extra
space when a single argument is given.

[1] https://lists.gnu.org/archive/html/bug-gnuastro/2022-03/msg00000.html