PSF subtract: adding --modelonly to not subtract the PSF
Until now, the script in charge of making the subtraction of the PSF was
not able to provide the modeled PSF (for each star). However, in some
situations it is good to be able to do it. For example if the modeling of
the stars within one image can be done independently (because the flux of
the stars don't affect each other). In this case, it would be worth to be
able to model all stars in paralel to speed up the process.

With this commit, a new option has been added in order to obtain the
scattered light field instead the default subtraction of the PSF. It adds
more flexibility and power when modeling the stars with the PSF.