Book: edited extended PSF overview and psf-select-stars sections
Until now, after starting work on the extended PSF tutorial, we hadn't
updated these sections in the actual documentation of the installed scripts
section! However, the scripts did undergo some evolution while we completed
the tutorial.

With this commit, these two sections were edited to be more clear, and
reflect all the changes that were made. For example, all the extra options
of the one-line examples of the 'astscript-psf-select-stars' script are no
longer necessary and would only scare the readers! So a single and clean
example (from the tutorial!) is placed there.

Also, I noticed that the description of the options were only
minimal/technical, and would force a new user to go into the script to
understand what that option means! So with this commit, the explanation has
been greatly expanded, giving more examples and more explanation.

Also, generally, when an option takes two numbers (like magnitude range),
the default format should be 'FLT,FLT', not 'INT[,INT[,...]]'. The latter
format says that any number of integer values are acceptable (which is
wrong).