DS9 script: fixing bug related to the input file variable name
Until this commit, variable name of the input file was wrongly used. It was
'input' while it should be 'inputs'. This problem cames from commit
"9947331d Installed scripts: not calling bash, several stylistic
improvements", when the variable 'inputs' was set but the old 'input' were
not appropriately replaced.

With this commit, all 'input' variables have been replaced by 'inputs' and
this problem has been fixed.