Installation: SAO DS9 and TOPCAT defined as optional dependencies
Until now, the 'astscript-fits-view' and 'astscript-ds9-region' would
assume that SAO DS9 already exists on the host operating
system. 'astscript-fits-view' would also assume that TOPCAT exists for
viewing tables. However, at configure time and in the description of
optional dependencies, there was no mention of DS9 or TOPCAT.

Also, in the extended PSF tutorial, the first time 'astscript-fits-view'
was run on a table, there was no mention that the user has to have TOPCAT
on their system.

With this commit, checks have been added in the configuration script so if
the user doesn't have them, they are warned/told about this. Also, an item
has been added for each in the optional dependencies and the commands to
install the dependencies from package managers now also include these
two.

Also, in the extended PSF tutorial, TOPCAT has been mentioned upon the
first usage of 'astscript-fits-view' on a table.

This issue was raised by Aaron Watkins.

In the process, I also noticed that some of the menus of the extended PSF
tutorial don't have any short description. So one was added for them.