astscript-sort-by-night: properly ordered input argument names
Until now, when given multiple files, through the 'inputs="$1 $inputs"'
expression, this script would effectively reverse their order for later
steps! Also, if there was a single argument, it would put an extra SPACE
before it.

With this commit, the standard check (that was recently added to other
scripts) has been added to make sure that the order is correct, and that
there isn't any extra SPACE before a single input.