PSF-scripts: added Infante-sainz et al. 2020 to --cite
Until now, when called with '--cite', all the extended PSF scripts would
simply call 'astfits --cite'! However, the method we use in these scripts
has been published in Raul's 2020 paper!

With this commit, a new 'print_citation' function has been added to all the
scripts. It will print the BibTeX information for Raul's paper
(https://ui.adsabs.harvard.edu/abs/2020MNRAS.491.5317I), as well as the
NoiseChisel paper (for Gnuastro and the tools used in the script in
general).

In the process, I also noticed that the common acknowledgement script for
the programs also doesn't reflect my (and Raul's) updated grant at CEFCA!
So the general statement was also edited.