PSF-scripts: make check tests now include tmpdir and output name
Until now, in the recipes for 'make check' tests, no temporary directory or
outputname was defined for them! This could potentially cause conflicts and
make things hard to debug.

With this commit, to be more clear, the test script for each PSF script now
has a '--tmpdir' and '--output' option that use the name of the script as
reference. This will help in debugging in case of errors.

Also, I noticed that the indentation of some of the script wasn't correct
(the options of the next lines should start after the '$execname' component
(since this is the actual program being run!): the '$check_with_program' is
usually empty (and will be replaced with 'valgrind' when the user
configures Gnuastro with '--enable-check-with-valgrind'.