Library (arithmetic): operators properly account for empty datasets
Until now, all the arithmetic functions in the library (and the
column-specific operators of Table's column arithmetic!), implicitly
assumed that an actual array exists in the input dataset. However,
especially when dealing with Column-arithmetic in Table, this is not always
the case: it can happen that the input table is from a previous automatic
command where no rows were selected!

In such cases, Table would abort with a "Segmentation fault (core dumped)",
which is not good and not informative!

With this commit a check has been added at the start of all the arithmetic
library and column-specific functions that will properly deal with empty
datasets.

This bug was reported by Zohreh Ghaffari.

This fixes bug #62204.