Book: spell check, updated THANKS in book, prepare for v.0.17
Until now, we were busy developing Gnuastro since version 0.16. However, we
are now ready to go to Gnuastro 0.17.

With this commit, following the checklist of 'doc/release-checklist.txt', a
spellcheck was run on the newly added parts of the book and the list of new
people who had helped in Gnuastro was updated in the book also (until not
it was just in 'THANKS').

Also, while doing a test packaging for Debian, I noticed that Lintian is
warning about the very long line in the new '--cite' options of the PSF
scripts. Therefore, it was broken into a multi-line paragraph.