MakeProfiles: unified function for reading input columns
Until now, when MakeProfiles was given an input from the standard input and
the user requested a 3D catalog, MakeProfiles would write the final output,
but abort with a segmentation fault.

This happened because the inputs for 2D and 3D inputs (which just have a
different number of columns) are done in separate functions (because the
number and meaning of the columns change).

We didn't confront this issue in 2D outputs, because this issue (of not
having a file name for the input when its from standard input) was solved
there. However, we had forgot to fix it in the same part of the 3D
function.

With this commit, to avoid having repetative code, a single low-level
function has been defined for only reading the desired columns and both
functions simply use that (after they have defined their necessary
columns). In this way, the bug has been fixed in 3D inputs also, and any
future bug or added feature just has to be implemented in one place.

Also, while updating this bug notice in the 'NEWS' file, I found some minor
missed things within the NEWS component of the 0.17 release that have been
fixed now.

This bug was found by Irene Pintos Castro.

This fixes bug #62216.