psf-stamp: now works with stars coordinate that are not in image
Until now, when a coordinate was not in the image, the 'astscrop' step for
finding the segment label (crop with very small size and crop center of the
'CLUMPS' and 'OBJECTS' main stars for masking the extra objects), the
script would abort. This happened because the crop width did not overlap
with input image. On the other hand in the normalizatin step, when all the
values within the requested radial range were blank, 'aststatistics' would
abort with an error and script would stop.

With this commit, an impossible label was given for the clump and object
labels, which are outside of the image and based on the artifical label all
extra objects will be masked. For solving the problem in the normalization
step, we manually set the value to NaN so the whole image becomes NaN (and
is not used when making a stack).

This fixes bug #62220.