Convolve: kernel is read before its sanity checks in PSF-matching
Until now, when requested to do PSF-matching, Convolve would crash with a
segmentation fault. The issue can be reproduced with the commands below.

    astmkprof --kernel=moffat,5,7,5 --oversample=1 --output=wide.fits
    astmkprof --kernel=gaussian,3,5 --oversample=1 --output=sharp.fits
    astconvolve wide.fits --kernel=sharp.fits --makekernel=10 \
                --output=matched-psf.fits

This happened because the check for the dimensionality of the kernel was
done before actually reading it!

With this commit, the issue has been addressed by bringing up the step to
read the kernel before the part where we check its dimensionality. Also the
documentation of the '--makekernel' option was slightly edited to be more
clear. To help catch PSF-matching crashes, a PSF-matchin test has been
added in the 'make check' phase.

This bug was found with the help of Raul Infante-Sainz.

This fixes bug #62250.