radial-profile: removes radii that are larger than requested rmax
Until now, due to the position angle and axis ratio, it was possible that
the output radial profile has one extra radius and we didn't have any check
to avoid this situation.

With this commit, the output catalog is checked so the final output does
not have radii larger than what user asked for.

This issue originates from MakeProfiles that built the labeled image for
measurement: it could happen that at certain ellipticities and position
angles, the profile would contain one or two pixels with a radius on the
limit due to floating-point errors. That pixel would cause this problem. So
the easiest solution was to limit the table in the end.