Library (arithmetic): more clear warning on integer sanity check
Until now, the warning message that is printed for binary operators (when
they are given integers of the same width as input), would simply say
"first/second popped operand". By "first" we mean the operand that is
closer to the operator. For example, in "2 3 gt", the first popped operand
is '3' and the second is '2'. But this could be confusing for some users
(who may confuse with the infix notation).

With this commit, some extra explanation is given in the warning message to
help clarify which one is "first" and which one is "second".

This was done with the help of Pedram Ashofte Ardakani.