Library (arithmetic.c): ra-to-degree checks for non-string inputs
Until now, the two 'ra-to-degree' or 'deg-to-degree' operators didn't do a
sanity check on the types of their inputs. As a result, they would give a
segmentation fault when not properly used (which is very bad and can cause
a lot of confusion for the user). This happened in the following two
scenarios:

    astarithmetic 16 ra-to-degree
    echo 16 | asttable -c'arith $1 ra-to-degree'
    echo 16h0m0 | asttable -c'arith $1 ra-to-degree'

In the third case, its because table automatically converts sexagesimal
formats to degrees in floating point (default type) columns.

With this commit, before this conversion is done with these operators, the
program will check if the input is string or not. If not, it will abort the
program and print an informative message on the cause and solution, to help
the user understand and fix the problem.

This bug was reported by Pedram Ashofte Ardakani.

This fixes bug #62253.